refactor: PostCompleteComponent로 분리
